requirements.txt
mypy==0.931
pylint==2.12.2
psycopg2-binary==2.9.3
Flask==2.1.2
